fix: harden blog integration after PR #90 production 500s
Split client-safe date formatting from MDX tooling, guard What's New
loading in root layout, trace content/blog into serverless bundles, and
align blog h2/h3 with display typography.
