+++ content/blog/2026-06-11-blog-launch.mdx
@@ -7,11 +7,11 @@ tags: [release]
 category: releases
 ---
 
-We are excited to announe the launch of the X-ray Atlas blog. This will be the primary place for announcements of new features, bug fixes, other updates to the platform, and an insigth into the development philosophy and process.
+We are excited to announce the launch of the X-ray Atlas blog. This will be the primary place for announcements of new features, bug fixes, other updates to the platform, and an insight into the development philosophy and process.
 
 # The Blog Platform and our Stakeholders
 
-I am comfortable saying that the stakeholders for this platform have spread beyond what we initially envisioned. We now have interested parties from various synchrotron facilities, beamline scientists, industry researchers, academic institutions, and more; each with their own level of interest and involvement in the platform. Because of this, we need to imporve our ability to communicate with these stakeholders at the level each group wants to engage at.
+I am comfortable saying that the stakeholders for this platform have spread beyond what we initially envisioned. We now have interested parties from various synchrotron facilities, beamline scientists, industry researchers, academic institutions, and more; each with their own level of interest and involvement in the platform. Because of this, we need to improve our ability to communicate with these stakeholders at the level each group wants to engage at.
 
 ## RSS Feed Subscription
